pyproject.toml
[project]
name = "headline-reactor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.8",
  "pytesseract>=0.3.10",
  "pillow>=10.3",
  "mss>=9.0",
  "pygetwindow>=0.0.9",
  "pywin32>=306; platform_system=='Windows'",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "rapidfuzz>=3.9",
  "openai>=1.40.0",        # Responses API (official Python SDK)
]

[project.optional-dependencies]
audio = ["sounddevice>=0.4"]

[project.scripts]
headline-reactor = "headline_reactor.cli:app"

